Bump the schema version to 3
This change just copies version 2 of the schema to a new version 3 to pave
the way for the removal of the actions.

Because we have not yet defined how version 3 will look like, the migration
path from 2 to 3 is not implemented and thus some tests are now broken (as
expected failures).